# [SPARK-25947][SQL] Reduce memory usage in ShuffleExchangeExec by selecting only the sort columns